+++ tests/test_cluster.cpp
@@ -8,8 +8,12 @@ using namespace std;
 
 class Cluster2 : public Cluster {
 public:
-    int64_t dist_to_box(coo z) { return max(int64_t(0), min({z.x - (ul.x + w - 1), z.y - (ul.y + w - 1), ul.x - z.x, ul.y - z.y})); }
-    int64_t dist_to_out(coo z) { return max(int64_t(0), min({z.x - ul.x + 1, z.y - ul.y + 1, (ul.x + w) - z.x, (ul.y + w) - z.y})); }
+    int64_t dist_to_box(coo z) {
+        return max(int64_t(0), min({z.x - (ul.x + int64_t(w) - 1), z.y - (ul.y + int64_t(w) - 1), ul.x - z.x, ul.y - z.y}));
+    }
+    int64_t dist_to_out(coo z) {
+        return max(int64_t(0), min({z.x - ul.x + 1, z.y - ul.y + 1, (ul.x + int64_t(w)) - z.x, (ul.y + int64_t(w)) - z.y}));
+    }
 
     int64_t inrad_go(coo z) {
         if (np == w * w) return dist_to_out(z);
